% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufo_functions.R
\name{count_sightings_by_day}
\alias{count_sightings_by_day}
\title{Count Sightings by Day of Week}
\usage{
count_sightings_by_day(data)
}
\arguments{
\item{data}{A data frame of UFO sightings.}
}
\value{
A data frame summarizing the number of sightings per day of the week.
}
\description{
This function counts the number of UFO sightings by day of the week.
}
