% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufo_functions.R
\name{count_sightings_by_month}
\alias{count_sightings_by_month}
\title{Count Sightings by Month}
\usage{
count_sightings_by_month(data)
}
\arguments{
\item{data}{A data frame of UFO sightings.}
}
\value{
A data frame summarizing the number of sightings per month.
}
\description{
This function counts the number of UFO sightings by month.
}
