% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufo_functions.R
\name{plot_sightings_by_shape}
\alias{plot_sightings_by_shape}
\title{Plot Sightings by Shape}
\usage{
plot_sightings_by_shape(data)
}
\arguments{
\item{data}{A data frame of UFO sightings.}
}
\value{
A ggplot object showing UFO sightings by shape.
}
\description{
This function creates a bar plot of UFO sightings by shape.
}
