% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufo_functions.R
\name{plot_sightings_over_time}
\alias{plot_sightings_over_time}
\title{Plot UFO Sightings Over Time}
\usage{
plot_sightings_over_time(data)
}
\arguments{
\item{data}{A data frame of yearly UFO sightings counts.}
}
\value{
A ggplot object showing UFO sightings over time.
}
\description{
This function creates a line plot of UFO sightings over the years.
}
