% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ufo_functions.R
\name{summarize_sightings_by_shape}
\alias{summarize_sightings_by_shape}
\title{Summarize UFO Sightings by Shape}
\usage{
summarize_sightings_by_shape(data)
}
\arguments{
\item{data}{A data frame of UFO sightings.}
}
\value{
A data frame summarizing the number of sightings per shape.
}
\description{
This function takes a dataset of UFO sightings and returns a summary
of sightings by shape.
}
